The unittest program has been updated.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/branches/RELENG_1_0@1981 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/rtm/tests/PeriodicExecutionContext/PeriodicExecutionContextTests.cpp
@@ -715,7 +715,7 @@ namespace PeriodicExecutionContext
 			
       RTC::ExecutionContextProfile* ecp;
       ecp = ec->get_profile();
-      CPPUNIT_ASSERT_EQUAL(CORBA::Double(0), ecp->rate);
+      CPPUNIT_ASSERT_EQUAL(CORBA::Double(1000000), ecp->rate);
 
       // set_rate()を呼出す
       CPPUNIT_ASSERT_EQUAL(RTC::RTC_OK, ec->set_rate(CORBA::Double(1.0)));
